More tests for Perk
